Great leaders are willing to sacrifice the numbers to save the people. Poor leaders sacrifice the people to save the numbers

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -45,5 +45,5 @@ int _atoi(char *s)
 
 	result = sign * result;
 
-	return (result);
+	return (int)(result);
 }
